// src/constants.d.ts
export declare const TEST_IMAGE_REF_ID = "image-9d014da5b91534cffd92822c7236916241f73369-2400x1600-png";
export declare const DEFAULT_IMAGE: {
    _type: string;
    asset: {
        _ref: string;
        _type: string;
    };
};
export declare const BASE_URL = "https://cdn.sanity.io/images/7ehtrhwv/production/9d014da5b91534cffd92822c7236916241f73369-2400x1600.png";